src/main/java/org/escuelaing/eci/repository/preference/Preference.java
package org.escuelaing.eci.repository.preference;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

import java.io.Serializable;
import java.util.Date;
import java.util.Objects;

@Document(collection = "Preference_collection")
public class Preference  implements Serializable{

    private static final long serialVersionUID = 1L;

    @Id
    private String id;
    private String foodType;
    private Integer value;

    

    public Preference(String id, String foodType, Integer value) {
        this.id = id;
        this.foodType = foodType;
        this.value = value;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getFoodType() {
        return foodType;
    }

    public void setFoodType(String foodType) {
        this.foodType = foodType;
    }

    public Integer getValue() {
        return value;
    }

    public void setValue(Integer value) {
        this.value = value;
    }

    @Override
    public String toString() {
        return "Preference{" +
                "id='" + id + '\'' +
                ", foodType='" + foodType + '\'' +
                ", value=" + value +
                '}';
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Preference that = (Preference) o;
        return Objects.equals(id, that.id) &&
                Objects.equals(foodType, that.foodType) &&
                Objects.equals(value, that.value);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, foodType, value);
    }
    
}
